Resolve #5 Implement CRUD Actions for Users
This commit implements CRUD actions for users with appropriate authorization and responses for user management following the pattern at https://github.com/plataformatec/devise/wiki/How-To:-Manage-users-through-a-CRUD-interface.

We move devise authorization paths to the ‘/my’ namespace for now.

We also make a minor fix to the DevelopmentsController to make it properly comport with the Pundit authorization patterns.