Resolve #4 Implement Moderation Feature
This implements the moderation feature by creating an Edit model that can accept proposed edits in a POST request. The proposed edits can then be approved by a superuser via a PATCH request. Changing the status to approved will trigger an after_save hook that will apply the edit to the development.