Resolve #30 Return Points with Development Data
This creates two classes of paramaters that modify the data retured from the developments index (GET /developments) endpoint. By passing a “trunc” paramater we return only a subset of the columns so that we can minimize the data loaded into the browser. We can also submit latitude and longitude parameters to define a boundig box to return points.