Export Shapefile
Provides functionality to export a shapefile by calling pgsql2shp utility in the command line. Still needs a function to zip up the constituent parts of the shapefile and return the entire thing as a zip to the user.

Also the relevant test is failing because the test database appears empty when it attempts to run the command line utility.