Fix shapefile download
This should allow a client to download a zip of a shapefile now. The issue was that we were deleting the .zip before the client could download the whole thing. By commenting it out and removing the filename argument the download works. Also rails registers zip as a file type by default so we do not need to register that.