Fix Response for Password Reset
This returns the correct kind of response for Ember using the advice at https://stackoverflow.com/questions/4800219/how-to-set-the-default-format-for-a-route-in-rails/41463065#41463065. As a general matter it appears Rails will use the extension of the path (i.e. test.json) in order to determine what to return in the repond_to block, and ignore the Accepts header. If there is no extension it returns HTML by default. It appears that JSONAPI requests subvert this issue by some kind of override in the ActiveModelSerializers gem that I did not find.