Fix pg_restore not working on local database
This fixes the pg_restore to work properly on a local machine. The first change we made was to add the -a flag to import only data and not the entire table infomration. This is because we create the table in the earlier part of the migration.

The second change is we fall back to using the system user for the username when we use pg_restore because if you do not define a user for pg_restore it does not work. On most systems there is a Postgres user role that corresponds with the system user so this should work for 80% of users. If it does not work then the user should set the database information in POSTGRES_USER.

Notes are also added to the README.md about setting these variables.