Fix for Active Model Serializers
Active Model Serializers was not serializing the parcel model correctly and so this fixes this by bumping the version to fix whatever bug was preventing it from occuring. Also we had to set the primary key for the parcel model to properly return an id element to present the parcel on the page in ember.