Move Database Setup Tasks to Seeds
Since we need our foreign data wrappers to be setup when we run our test suite and development databases, the correct home for it is in seeds.rb where they will be invoked on database creation in all environments. As a result of this improvement we can then remove some setup steps from our README.md.