Resolve #281 Allow Closing Splash on iPhone
Content, it’s fun. We write. We share. Unfortunately our ideas can be bigger than the divs we planned to contain them. Users then lose the ability to close a modal on their iPhones. They cannot see our site. Sad! This fix removes content, for now, to allow most users to access the site. When we have time and resources we can bring this back.