Fix Leaky Flag Policy
The update? policy would always return true becuase [:flag] si always true. This functionally worked when we wanted anyone to update their flags, but meant that the update action was wide open. Not so great! Here were refactor the controller to use an authorization scheme based on the presence of a flag parameter. This is likely the best we can do for now, it’s not clear that we can give our Policy instance knowledge of the params to break them up.