Resolve Invalid Date in Safari + Firefox
We learned that in Safari and Firefox the created_at and updated_at dates were displaying “Invalid Date” instead of the actual date. This turned out to be because the date provided by the Rails API was not properly formatted as ISO 8601. This commit updates the serializers to output ISO8601 formatted dates so we can now see the dates in these two browsers.