Eliminate ember-string-helpers dependency
The ember-string-helpers dependency used a deprecated method of summoning help from node-modules. We only used a limited portion of this rather small library, so we can easily refactor the string helpers into our own internal helpers. This commit stubs out the internal helpers and eliminates the dependency, leaving the app open to finishing the helper in a future commit.