Remove Redundant default method call
Currently our parcel outlines are not appearing on the map in this version. We attempt to remedy this issue by noting from the documentation at https://www.npmjs.com/package/@turf/boolean-point-in-polygon that they do not call .default, and in our JavaScript console the app makes a call to `.default.default`. Hopefully removing this method call fixes the parcel display.

Co-authored-by: Annabelle Thomas Taylor <ataylor@mapc.org>