Replace call to CARTO with static array of municipalities
In order to connect a member of city/town government account with a municipality, we have a drop-down list of MA municipalities. Since we moved away from CARTO, our old request not longer worked. Given that the names do not change frequently, we decided to instead pull from a static list and save ourselves a GET request.